move config around and create a remote configuration file
we'll use this to figure out which remotes we can ignore/what button
presses are legit for what kinds of remote, etc.
