+++ src/caseta/mod.rs
@@ -1,6 +1,3 @@
 pub mod connection;
 pub mod message;
-mod scene;
 pub mod remote;
-
-pub use scene::*;
