+++ src/config/mod.rs
@@ -1,0 +1,3 @@
+pub mod caseta_hub_configuration;
+mod scene;
+pub mod caseta_remote;
